ingestion_pipeline/requirements.txt
fastapi
uvicorn
dotenv
python-docx
PyPDF2
python-multipart
